Add commercial migrations and rename tables to consumer, make sure migrations work forwards and backwards with --force [#1366]

+++ lib/generators/templates/migration_veda_credit_request.rb
@@ -1,18 +1,20 @@
 class CreateVedaCreditRequest < ActiveRecord::Migration
   def self.up
-    create_table :veda_credit_requests do |t|
-      t.integer :application_id
-      t.text :xml
-      t.text :access
-      t.text :service
-      t.text :entity
-      t.text :bureau_reference
-      t.text :enquiry
-      t.timestamps
+    unless ActiveRecord::Base.connection.table_exists? 'veda_credit_requests'
+      create_table :veda_credit_requests do |t|
+        t.integer :application_id
+        t.text :xml
+        t.text :access
+        t.text :service
+        t.text :entity
+        t.text :bureau_reference
+        t.text :enquiry
+        t.timestamps
+      end
     end
   end
   
   def self.down
-    drop_table :veda_credit_requests
+    drop_table :veda_credit_requests if ActiveRecord::Base.connection.table_exists? 'veda_credit_requests'
   end
 end
